Refactor barsukas API routes into package modules
